YoloRealTime_Detection.py

import os,cv2,time,threading,sys,json,numpy as np
from pathlib import Path
from datetime import datetime
from typing import Dict,List,Optional,Tuple
import tkinter as tk
from tkinter import ttk,filedialog,messagebox
from PIL import Image,ImageTk
try:
    from pygrabber.dshow_graph import FilterGraph
except ImportError:
    FilterGraph=None

# Import YOLODetector from detection.detector
from detection.detector import YOLODetector
from gui.main_gui import YOLOGui

if __name__ == "__main__":
    YOLOGui().run()