requirements.txt
ultralytics
torch
opencv-python
numpy
Pillow
pygrabber